app/requirements.txt
alembic==1.16.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.4
beautifulsoup4==4.13.4
web3==7.12.0
click==8.1.8